requirements.txt
streamlit
llama-cpp-python
torch
transformers
sentencepiece
accelerate
